// src/contracts/generated/Version.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 协议版本（握手交换用）。
 */
export type Version = { 
/**
 * 主版本。**必须相同**。
 */
major: number, 
/**
 * 次版本。**服务端 ≥ 客户端**。
 */
minor: number, 
/**
 * 补丁版本（不参与兼容判定）。
 */
patch: number, };